# swagger-spec.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}},"tags":["App"]}},"/user/apk":{"get":{"description":"读取 package/release.json，比对 ?version；一致返回“已是最新”；不一致返回储存桶下载 URL、最新版本、日期与更新内容","operationId":"UserController_apk","parameters":[{"name":"version","required":false,"in":"query","description":"客户端当前版本","schema":{"type":"string"}}],"responses":{"200":{"description":"返回最新版本信息或储存桶下载地址","content":{"application/json":{"schema":{"type":"object","properties":{"isLatest":{"type":"boolean","example":false},"latestVersion":{"type":"string","example":"2.9.0"},"date":{"type":"string","example":"2025-10-21"},"content":{"type":"array","items":{"type":"string"}},"url":{"type":"string","example":"https://dx.jcw2.cn/suixinji-2.9.0_anzhuo.apk"},"message":{"type":"string","example":"已经是最新版本"}}}}}}},"summary":"获取最新 APK（返回储存桶 URL）","tags":["User"]}},"/user/register":{"post":{"description":"传昵称和密码，返回分配的账号","operationId":"UserController_register","parameters":[],"requestBody":{"required":true,"description":"注册请求体","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"200":{"description":"注册成功","content":{"application/json":{"schema":{"type":"object","properties":{"account":{"type":"integer","example":10000000}}}}}}},"summary":"注册用户","tags":["User"]}},"/user/login":{"post":{"description":"使用账号与密码登录，返回永不过期 token 和用户信息（不含密码）","operationId":"UserController_login","parameters":[],"requestBody":{"required":true,"description":"登录请求体","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginUserDto"}}}},"responses":{"200":{"description":"登录成功","content":{"application/json":{"schema":{"type":"object","properties":{"token":{"type":"string","example":"eyJhbGciOi..."},"id":{"type":"integer","example":1},"account":{"type":"integer","example":10000000},"nickname":{"type":"string","example":"张三"},"avatar":{"type":"string","nullable":true,"example":"https://example.com/avatar.png"},"isVip":{"type":"boolean","example":false},"registerTime":{"type":"string","format":"date-time","example":"2025-10-17T08:00:00.000Z"}}}}}}},"summary":"登录","tags":["User"]}},"/user":{"post":{"operationId":"UserController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"200":{"description":"创建成功"}},"summary":"创建用户（同注册）","tags":["User"]},"get":{"operationId":"UserController_findAll","parameters":[],"responses":{"200":{"description":"返回用户列表"}},"summary":"获取所有用户","tags":["User"]}},"/user/{id}":{"get":{"operationId":"UserController_findOne","parameters":[{"name":"id","required":true,"in":"path","description":"用户ID","schema":{"example":1,"type":"number"}}],"responses":{"200":{"description":"返回用户详情"}},"summary":"获取指定用户","tags":["User"]},"patch":{"operationId":"UserController_update","parameters":[{"name":"id","required":true,"in":"path","description":"用户ID","schema":{"example":1,"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":"更新成功"}},"summary":"更新指定用户","tags":["User"]},"delete":{"operationId":"UserController_remove","parameters":[{"name":"id","required":true,"in":"path","description":"用户ID","schema":{"example":1,"type":"number"}}],"responses":{"200":{"description":"删除成功"}},"summary":"删除指定用户","tags":["User"]}},"/review/{parentId}/reply":{"post":{"description":"仅允许回复根评价（不可对回复再回复），只需消息内容","operationId":"ReviewController_reply","parameters":[{"name":"parentId","required":true,"in":"path","description":"父评价ID（必须为根评价）","schema":{"example":10,"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReplyReviewDto"}}}},"responses":{"200":{"description":"回复成功，返回回复对象"}},"security":[{"bearer":[]}],"summary":"回复评价（仅一级）","tags":["Review"]}},"/review":{"post":{"description":"无需消息ID，系统会将该评价的 id 作为 messageId","operationId":"ReviewController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateReviewDto"}}}},"responses":{"200":{"description":"创建成功，返回评价对象"}},"security":[{"bearer":[]}],"summary":"创建评价（根评价）","tags":["Review"]}},"/review/messages":{"get":{"description":"返回所有根评价及其直接回复","operationId":"ReviewController_findAllMessages","parameters":[],"responses":{"200":{"description":"查询成功"}},"security":[{"bearer":[]}],"summary":"获取全部消息（根评价线程）","tags":["Review"]}},"/review/message/{messageId}":{"get":{"description":"返回该消息的所有根评价及其直接回复","operationId":"ReviewController_findByMessage","parameters":[{"name":"messageId","required":true,"in":"path","description":"消息ID","schema":{"example":1001,"type":"number"}}],"responses":{"200":{"description":"查询成功"}},"security":[{"bearer":[]}],"summary":"按消息查询评价列表","tags":["Review"]},"delete":{"description":"按 messageId 删除根评价及其直接回复","operationId":"ReviewController_deleteMessage","parameters":[{"name":"messageId","required":true,"in":"path","description":"消息ID（根评价的 id）","schema":{"example":1001,"type":"number"}}],"responses":{"200":{"description":"删除成功"}},"security":[{"bearer":[]}],"summary":"删除消息（整条评价线程）","tags":["Review"]}},"/review/{id}":{"get":{"description":"包含该评价的直接回复","operationId":"ReviewController_findOne","parameters":[{"name":"id","required":true,"in":"path","description":"评价ID","schema":{"example":1,"type":"number"}}],"responses":{"200":{"description":"查询成功"}},"security":[{"bearer":[]}],"summary":"评价详情","tags":["Review"]}},"/export/excel":{"post":{"description":"将前端导出页面的数据发送到后端，后端生成 Excel 保存在项目根目录 exportFiles，并返回下载 URL。文件名规则：优先使用前端提供的 fileName；否则使用 {account}_{YYYY-MM-DD}.xlsx。","operationId":"ExportController_exportExcel","parameters":[],"requestBody":{"required":true,"description":"导出请求体","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateExportDto"}}}},"responses":{"200":{"description":"生成成功，返回下载地址","content":{"application/json":{"schema":{"type":"object","properties":{"filename":{"type":"string","example":"10000000_2025-10-26.xlsx"},"url":{"type":"string","example":"http://localhost:3000/export/file/10000000_2025-10-26.xlsx"}}}}}}},"summary":"生成 Excel 并返回下载地址","tags":["Export"]}},"/export/file/{filename}":{"get":{"operationId":"ExportController_downloadExcel","parameters":[{"name":"filename","required":true,"in":"path","description":"Excel 文件名","schema":{"example":"10000000_2025-10-26.xlsx","type":"string"}}],"responses":{"200":{"description":"返回 Excel 文件（二进制流）"}},"summary":"下载生成的 Excel 文件","tags":["Export"]}}},"info":{"title":"template-nest","description":"","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"CreateUserDto":{"type":"object","properties":{"nickname":{"type":"string","description":"昵称","minLength":2,"maxLength":64,"example":"张三"},"password":{"type":"string","description":"密码","minLength":6,"maxLength":128,"example":"secret123"}},"required":["nickname","password"]},"LoginUserDto":{"type":"object","properties":{"account":{"type":"number","description":"账号（从 10000000 起）","minimum":10000000,"example":10000000},"password":{"type":"string","description":"密码","minLength":6,"maxLength":128,"example":"secret123"}},"required":["account","password"]},"UpdateUserDto":{"type":"object","properties":{}},"ReplyReviewDto":{"type":"object","properties":{"content":{"type":"string","description":"回复内容","example":"同意你的看法","minLength":1,"maxLength":1000}},"required":["content"]},"CreateReviewDto":{"type":"object","properties":{"rating":{"type":"number","description":"评分(1-5)","example":5,"minimum":1,"maximum":5},"content":{"type":"string","description":"评价内容","example":"非常不错！","minLength":1,"maxLength":1000}},"required":["rating","content"]},"ExportRecordDto":{"type":"object","properties":{"date":{"type":"string","description":"日期","example":"2025-10-26"},"timestamp":{"type":"number","description":"时间戳","example":1761451200000},"amount":{"type":"number","description":"金额","example":35.2},"billType":{"type":"string","description":"类型","example":"expense","enum":["expense","income","transfer","reimbursement","refund"]},"category":{"type":"object","description":"分类对象","example":{"name":"奶茶","icon":"🧋"}},"note":{"type":"string","description":"备注","example":"午餐"},"paymentMethod":{"type":"string","description":"支付方式代码","example":"wechat"},"paymentDetail":{"type":"string","description":"支付详情（前端实际字段）","example":""},"paymentAccount":{"type":"object","description":"支付账户（支出/转账）","example":{"id":"account-1761299847431-g4s1wovei","name":"测试","icon":"yh","method":"cash"}},"incomeAccountInfo":{"type":"object","description":"收入账户信息（收入）","example":{"id":"account-yy","name":"储蓄卡","icon":"💳"}},"incomeAccount":{"type":"string","description":"收入账户（字符串，前端实际字段）","example":""},"targetAccountInfo":{"type":"object","description":"转账目标账户信息（内部转账）","example":{"id":"account-zz","name":"余额宝","icon":"💰"}},"reimbursementAccount":{"type":"object","description":"报销账户对象","example":{"id":"account-aa","name":"公司报销账户","icon":"🏢"}},"reimbursementAmount":{"type":"number","description":"已报销金额","example":20},"transferType":{"type":"string","description":"转账类型","example":"internal"},"targetAccount":{"type":"string","description":"转至账户名称（外部转账）","example":"好友张三"},"images":{"description":"图片列表（前端实际字段）","example":[],"type":"array","items":{"type":"string"}},"id":{"type":"number","description":"记录ID（前端实际字段）","example":1761463673091},"isImportant":{"type":"boolean","description":"是否重要（前端实际字段）","example":false},"printGroupId":{"type":"string","description":"打印分组ID（前端实际字段）","example":"grp_1761463675184_32149"}},"required":["amount","billType"]},"ExportAccountDto":{"type":"object","properties":{"name":{"type":"string","description":"账户名称","example":"微信钱包"},"type":{"type":"string","description":"账户类型（前端实际字段）","example":"cash"},"id":{"type":"string","description":"账户ID（前端实际字段）","example":"account-1761299847431-g4s1wovei"},"categoryId":{"type":"string","description":"分类ID","example":"category-1"},"balance":{"type":"number","description":"余额","example":120.5},"icon":{"type":"string","description":"图标","example":"yh"},"color":{"type":"string","description":"颜色","example":"#4CAF50"},"description":{"type":"string","description":"描述","example":"主要微信账户"},"isDefault":{"type":"boolean","description":"默认账户","example":false},"excludeFromTotal":{"type":"boolean","description":"不计入总资产","example":false},"isCreditAccount":{"type":"boolean","description":"是否信贷账户","example":false},"createdAt":{"type":"number","description":"创建时间戳","example":1761299847431},"updatedAt":{"type":"number","description":"更新时间戳","example":1761463673091}},"required":["name","balance"]},"ExportCategoryDto":{"type":"object","properties":{"id":{"type":"string","description":"分类ID","example":"category-123"},"name":{"type":"string","description":"分类名称","example":"餐饮"}},"required":["id","name"]},"CreateExportDto":{"type":"object","properties":{"account":{"type":"number","description":"用户ID（用于文件名）","example":10000000},"fileName":{"type":"string","description":"导出文件名（可选，默认 {account}_{YYYY-MM-DD}.xlsx）","example":"记账数据_2025-10-26.xlsx"},"records":{"description":"记录列表","type":"array","items":{"$ref":"#/components/schemas/ExportRecordDto"}},"accounts":{"description":"账户列表","type":"array","items":{"$ref":"#/components/schemas/ExportAccountDto"}},"categories":{"description":"分类映射（可选）","type":"array","items":{"$ref":"#/components/schemas/ExportCategoryDto"}}},"required":["account","records","accounts"]}}},"externalDocs":{"description":"api-json","url":"http://localhost:3000/api-json"}}